packages/viz/vizualization_py/requirements.txt
streamlit
streamlit-extras
python-dotenv
leafmap
mysqlclient
pymysql
Django
SQLAlchemy
flytekitplugins-bacalhau
bacalhau-sdk
eth-ape>=0.7.11 ## optional: in order to integrate the read operatioàns directly in the frontend